Split the Android activity to smaller files
